# PERF: improves performance and memory usage of DataFrame.duplicated