@shoyer 2015-02-03
Nice! I noticed the other day that duplicated did not use the klib hashtable, which seemed strange, but I didn't have the time to dig into it.

Can you safely get ride of `pandas.lib.duplicated` with this change?