@jreback 2015-02-24
I meant someting like this: https://github.com/jreback/pandas/commit/923e35c2ce9c3fd0c5dac5c6f281d470e98072ff