@behzadnouri 2015-02-25
I still feel the current PR is cleaner and avoid the unnecessary type check/manipulation in [here](https://github.com/pydata/pandas/blob/1d94dd2ada7d0af57c245b083d26d60b8649a04e/pandas/core/algorithms.py#L128) and [here](https://github.com/pydata/pandas/blob/1d94dd2ada7d0af57c245b083d26d60b8649a04e/pandas/core/algorithms.py#L164); but, if you like to make these changes, plz do.