@jreback 2015-03-03
merged via https://github.com/pydata/pandas/commit/7da9178dad2c4b000c2446e5cc58ff0613a631c1

thanks @behzadnouri